requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
numpy==1.19.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
whitenoise==5.2.0
